+++ app/components/map/services/map-service.js
@@ -125,13 +125,11 @@ angular.module('map')
           if (bounds instanceof LeafletService.LatLngBounds) {
             service._map.fitBounds.apply(service._map, arguments);
           }
-          else if (bounds.hasOwnProperty('south')
-            && bounds.hasOwnProperty('north')
-            && bounds.hasOwnProperty('east')
-            && bounds.hasOwnProperty('west')) {
+          else if (bounds.hasOwnProperty('_southWest')
+            && bounds.hasOwnProperty('_northEast')) {
             service._map.fitBounds(L.latLngBounds(
-              L.latLng(bounds.south, bounds.east),
-              L.latLng(bounds.north, bounds.west)));
+              L.latLng(bounds._southWest),
+              L.latLng(bounds._northEast)));
           }
         }
       },
@@ -361,7 +359,8 @@ angular.module('map')
         map.on('click',     function (e) { conditionalApply(cbs.onClick, e); });
         map.on('movestart', function (e) { conditionalApply(cbs.onMoveStart, e); });
         map.on('mousemove', function (e) { conditionalApply(cbs.onMouseMove, e); });
-        map.on('moveend',   function (e) { conditionalApply(cbs.onMoveEnd, e); });
+        map.on('moveend', function (e) { conditionalApply(cbs.onMoveEnd, e); });
+        map.on('zoomend', function (e) { conditionalApply(cbs.onZoomEnd, e); });
       },
 
       /**